+++ contrib/mac/framework/Julia.h
@@ -1,0 +1,4 @@
+#include <Julia/julia/julia.h>
+#include <Julia/julia/julia_threads.h>
+#include <Julia/julia/julia_version.h>
+#include <Julia/julia/uv.h>