+++ stdlib/LinearAlgebra/test/bunchkaufman.jl
@@ -72,6 +72,21 @@ bimg  = randn(n,2)/2
                 @test_throws ErrorException bc1.Z
                 @test_throws ArgumentError uplo == :L ? bc1.U : bc1.L
             end
+            # test Base.iterate
+            ref_objs = (bc1.D, uplo == :L ? bc1.L : bc1.U, bc1.p)
+            for (bki, bkobj) in enumerate(bc1)
+                @test bkobj == ref_objs[bki]
+            end
+            if eltya <: BlasFloat
+                @test convert(LinearAlgebra.BunchKaufman{eltya}, bc1) === bc1
+                @test convert(LinearAlgebra.Factorization{eltya}, bc1) === bc1
+                if eltya <: BlasReal
+                    @test convert(LinearAlgebra.Factorization{Float16}, bc1) == convert(LinearAlgebra.BunchKaufman{Float16}, bc1)
+                elseif eltya <: BlasComplex
+                    @test convert(LinearAlgebra.Factorization{ComplexF16}, bc1) == convert(LinearAlgebra.BunchKaufman{ComplexF16}, bc1)
+                end
+            end
+            @test Base.propertynames(bc1) == (:p, :P, :L, :U, :D)
         end
 
         @testset "$eltyb argument B" for eltyb in (Float32, Float64, ComplexF32, ComplexF64, Int)