+++ stdlib/LinearAlgebra/test/symmetric.jl
@@ -123,6 +123,8 @@ end
                         @test Hermitian(convert(Matrix{typ},aherm)) == convert(Hermitian{typ,Matrix{typ}},Hermitian(aherm))
                     end
                 end
+                @test Symmetric{eltya, Matrix{eltya}}(Symmetric(asym, :U)) === Symmetric(asym, :U)
+                @test Hermitian{eltya, Matrix{eltya}}(Hermitian(aherm, :U)) === Hermitian(aherm, :U)
             end
 
             @testset "issymmetric, ishermitian" begin
@@ -154,7 +156,7 @@ end
             @testset "transpose, adjoint" begin
                 S = Symmetric(asym)
                 H = Hermitian(aherm)
-                @test  transpose(S) === S == asym
+                @test transpose(S) === S == asym
                 @test adjoint(H) === H == aherm
                 if eltya <: Real
                     @test adjoint(S) === S == asym
@@ -163,6 +165,8 @@ end
                     @test adjoint(S) ==  Symmetric(conj(asym))
                     @test transpose(H) ==  Hermitian(copy(transpose(aherm)))
                 end
+                @test copy(adjoint(H)) == copy(aherm)
+                @test copy(transpose(S)) == copy(asym)
             end
 
             @testset "real, imag" begin
@@ -590,6 +594,13 @@ end
     @test_throws ArgumentError Symmetric(B) == Symmetric(Matrix(B))
     A[1,1] = 1; A[2,2] = 4
     @test Symmetric(B) == Symmetric(Matrix(B))
+end
+
+@testset "symmetric()/hermitian() for Numbers" begin
+    @test LinearAlgebra.symmetric(1, :U) == 1
+    @test LinearAlgebra.symmetric_type(Int) == Int
+    @test LinearAlgebra.hermitian(1, :U) == 1
+    @test LinearAlgebra.hermitian_type(Int) == Int
 end
 
 end # module TestSymmetric